+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/Discovery/InternalDiscoveryResult.cs
@@ -4,5 +4,7 @@ public record InternalDiscoveryResult(
     Reqnroll.VisualStudio.ReqnrollConnector.Models.StepDefinition[] StepDefinitions,
     Reqnroll.VisualStudio.ReqnrollConnector.Models.Hook[] Hooks,
     IDictionary<string, string> SourceFiles,
-    IDictionary<string, string> TypeNames
+    IDictionary<string, string> TypeNames,
+    string[] GenericBindingErrors,
+    string[] TypeLoadErrors
 );