+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Models/ConnectorResult.cs
@@ -7,6 +7,7 @@ public abstract class ConnectorResult
     public string ReqnrollVersion { get; set; }
     public string ErrorMessage { get; set; }
     public bool IsFailed => !string.IsNullOrWhiteSpace(ErrorMessage);
+    public string[] LogMessages { get; set; }
     public string[] Warnings { get; set; }
     public Dictionary<string, object> AnalyticsProperties { get; set; }
 }