+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support file scoped namespace declarations when generating code (#140)
 * Improved handling of scope tag expressions, hook and scope errors (#150)
 * Improved logging for binding discovery (#154)
+* Report generic binding errors and log type load errors of binding discovery (#157)
 
 ## Bug fixes: